Support looking up icons and masks in families by their resolution
This adds a new element_types submodule, which provides information
about the type of data stored in icon family elements with a certain
type code, as well as the resolution (in case of actual icon elements).
It also includes "quality" rankings for both icon and mask types,
indicating which type should be preferred if multiple types for the
same resolution are present in a family.

Some of this information was already available through the
Kaitai Struct-based parser, but because KS doesn't support proper
constants, the information could only be accessed after parsing the
element in question. This is not good for performance - the
type/resolution information should be available before parsing, so that
only the elements with the desired resolution need to be parsed.

Using this information, it's now possible to group icons in a family by
their resolution and look up the best available icon for each
resolution. If the best available icon doesn't include a mask, some
extra logic is needed to also look up the best available mask of the
same resolution and combine it with the maskless image. IconFamily now
stores icons grouped by resolution in this way and provides methods for
looking up and combining icons and masks.