Remove not very useful IconFamily.from_stream ensure_root parameter
It's actually not all that likely to find data in a format where
ensure_root=False would be useful, because such data would have to
include a header (just with a type code other than 'icns').
A tool/library that extracts raw data from icon family elements
probably wouldn't preserve the header when extracting the data, because
for example PNG and JPEG 2000 icon data can only be used standalone
*without* the icon family element header.

For now, if you want to parse raw icon family data that doesn't have a
proper 'icns' element header, you should either replace the type code
of the existing header (if any) with 'icns', or add the header (if it's
missing).